Added you died popup

+++ Assets/Scripts/Character/Player/PlayerUI/PlayerUIHudManager.cs
@@ -16,7 +16,7 @@ namespace TravserserProject
             staminaBar.gameObject.SetActive(false);
             staminaBar.gameObject.SetActive(true);
         }
-        public void SetNewHealthValue(float oldValue, float newValue)
+        public void SetNewHealthValue(int oldValue, int newValue)
         {
             healthBar.SetStat(Mathf.RoundToInt(newValue));
         }
